getting rid of unneeded years for un and allydir
the reason why I was having trouble before was because the DyadBuild
function matches based on the country years in the 'panel.rda'. So
while the country years in the 'panel.rda' dataset accurately catalogue
which countries are in existence and when, this wasn't mapping on well
in terms of which countries were UN members and when.

To get at that, I basically just threw in a conditional into the
DyadBuild function to match to country-years when the option is
specified.

Though theoretically this shouldn't be a problem for the directed ally
data, I did the same thing for it because of our variance issue.

Also last really small thing, I added a couple of line of code into the
setup.R file :

panel$cnameYear<-toupper(panel$cnameYear)
panel$cname<-toupper(panel$cname)

because for some reason German Democratic Republic is not capitalized.
If you have time, you might want to change whatever file you orignally
used to create the panel.rda file to fix that?